Core - extract UI components Core.UI - create new library for common UI functionality Core.UI.WPF - create new library for WPF UI functionality
