+++ SoluiNet.DevTools.Core.UI.WPF/UIElement/BrushDefinitionUserControl.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="SoluiNet.DevTools.Core.UIElement.BrushDefinitionUserControl"
+﻿<UserControl x:Class="SoluiNet.DevTools.Core.UI.WPF.UIElement.BrushDefinitionUserControl"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
