+++ SoluiNet.DevTools.Core.UI.WPF/UIElement/Editor/CompletionData.cs
@@ -2,7 +2,7 @@
 // Copyright (c) SoluiNet. All rights reserved.
 // </copyright>
 
-namespace SoluiNet.DevTools.Core.UI.UIElement.Editor
+namespace SoluiNet.DevTools.Core.UI.WPF.UIElement.Editor
 {
     using System;
     using System.Collections.Generic;
