+++ SoluiNet.DevTools.Core.UI.WPF/UIElement/ExtendedConfigurationUserControl.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="SoluiNet.DevTools.Core.UIElement.ExtendedConfigurationUserControl"
+﻿<UserControl x:Class="SoluiNet.DevTools.Core.UI.WPF.UIElement.ExtendedConfigurationUserControl"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
