+++ SoluiNet.DevTools.Core.UI/UIElement/ISoluiNetUIElement.cs
