+++ SoluiNet.DevTools.Core/XmlData/BrushDefinition.xsd
