+++ SoluiNet.DevTools.Core/XmlData/ExtendedConfiguration.xsd
