+++ SoluiNet.DevTools.Core/XmlData/generate_xsd.bat
@@ -1,0 +1,5 @@
+:: "%programfiles(x86)%\Microsoft SDKs\Windows\v10.0A\bin\NETFX 4.6.1 Tools\xsd.exe" /c /namespace:SoluiNet.DevTools.Core.XmlData /language:CS BrushDefinition.xsd
+:: "%programfiles(x86)%\Microsoft SDKs\Windows\v10.0A\bin\NETFX 4.7.2 Tools\xsd.exe" /c /namespace:SoluiNet.DevTools.Core.XmlData /language:CS BrushDefinition.xsd
+
+"%programfiles(x86)%\Microsoft SDKs\Windows\v10.0A\bin\NETFX 4.6.1 Tools\xsd.exe" /c /namespace:SoluiNet.DevTools.Core.XmlData /language:CS ExtendedConfiguration.xsd BrushDefinition.xsd
+"%programfiles(x86)%\Microsoft SDKs\Windows\v10.0A\bin\NETFX 4.7.2 Tools\xsd.exe" /c /namespace:SoluiNet.DevTools.Core.XmlData /language:CS ExtendedConfiguration.xsd BrushDefinition.xsd
