+++ SoluiNet.DevTools.UI.Sql/SqlUiElement.cs
@@ -34,7 +34,7 @@ namespace SoluiNet.DevTools.UI.Sql
     using SoluiNet.DevTools.Core.Tools.UI;
     using SoluiNet.DevTools.Core.UI;
     using SoluiNet.DevTools.Core.UI.UIElement;
-    using SoluiNet.DevTools.Core.UI.UIElement.Editor;
+    using SoluiNet.DevTools.Core.UI.WPF.UIElement.Editor;
 
     /// <summary>
     /// An UI element which can be used to provide database and SQL related functions.
