+++ SoluiNet.DevTools.UI/MainWindow.xaml
@@ -5,7 +5,7 @@
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:local="clr-namespace:SoluiNet.DevTools.UI"
         xmlns:avalonedit="http://icsharpcode.net/sharpdevelop/avalonedit"
-        xmlns:ui="clr-namespace:SoluiNet.DevTools.Core.UI.Window;assembly=SoluiNet.DevTools.Core"
+        xmlns:ui="clr-namespace:SoluiNet.DevTools.Core.UI.WPF.Window;assembly=SoluiNet.DevTools.Core.UI.WPF"
         xmlns:tb="http://www.hardcodet.net/taskbar"
         mc:Ignorable="d"
         Title="SoluiNet.DevTools" TitleFormatString=" {1} ({0}) - SoluiNet.DevTools" Height="450" Width="800">
