+++ SoluiNet.DevTools.UI/SoluiNet.DevTools.UI.csproj
@@ -96,6 +96,8 @@
     <EmbeddedResource Include="SQL.xshd" />
   </ItemGroup>
   <ItemGroup>
+    <ProjectReference Include="..\SoluiNet.DevTools.Core.UI.WPF\SoluiNet.DevTools.Core.UI.WPF.csproj" />
+    <ProjectReference Include="..\SoluiNet.DevTools.Core.UI\SoluiNet.DevTools.Core.UI.csproj" />
     <ProjectReference Include="..\SoluiNet.DevTools.Core\SoluiNet.DevTools.Core.csproj" />
   </ItemGroup>
   <ItemGroup>
