+++ SoluiNet.DevTools.sln
@@ -65,6 +65,10 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Install", "Install", "{F99C
 EndProject
 Project("{930C7802-8A8C-48F9-8165-68863BCCD9DD}") = "SoluiNet.DevTools.Install", "SoluiNet.DevTools.Install\SoluiNet.DevTools.Install.wixproj", "{040B16E4-BEC3-4429-920D-C0D206B1D6D9}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SoluiNet.DevTools.Core.UI", "SoluiNet.DevTools.Core.UI\SoluiNet.DevTools.Core.UI.csproj", "{01B12B16-26AB-41EF-9D2F-91894F0CE19A}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SoluiNet.DevTools.Core.UI.WPF", "SoluiNet.DevTools.Core.UI.WPF\SoluiNet.DevTools.Core.UI.WPF.csproj", "{E7AE24F6-004C-4CE3-86D5-58D7F5FE89CC}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -345,6 +349,30 @@ Global
 		{040B16E4-BEC3-4429-920D-C0D206B1D6D9}.Release|x64.ActiveCfg = Release|x86
 		{040B16E4-BEC3-4429-920D-C0D206B1D6D9}.Release|x86.ActiveCfg = Release|x86
 		{040B16E4-BEC3-4429-920D-C0D206B1D6D9}.Release|x86.Build.0 = Release|x86
+		{01B12B16-26AB-41EF-9D2F-91894F0CE19A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{01B12B16-26AB-41EF-9D2F-91894F0CE19A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{01B12B16-26AB-41EF-9D2F-91894F0CE19A}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{01B12B16-26AB-41EF-9D2F-91894F0CE19A}.Debug|x64.Build.0 = Debug|Any CPU
+		{01B12B16-26AB-41EF-9D2F-91894F0CE19A}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{01B12B16-26AB-41EF-9D2F-91894F0CE19A}.Debug|x86.Build.0 = Debug|Any CPU
+		{01B12B16-26AB-41EF-9D2F-91894F0CE19A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{01B12B16-26AB-41EF-9D2F-91894F0CE19A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{01B12B16-26AB-41EF-9D2F-91894F0CE19A}.Release|x64.ActiveCfg = Release|Any CPU
+		{01B12B16-26AB-41EF-9D2F-91894F0CE19A}.Release|x64.Build.0 = Release|Any CPU
+		{01B12B16-26AB-41EF-9D2F-91894F0CE19A}.Release|x86.ActiveCfg = Release|Any CPU
+		{01B12B16-26AB-41EF-9D2F-91894F0CE19A}.Release|x86.Build.0 = Release|Any CPU
+		{E7AE24F6-004C-4CE3-86D5-58D7F5FE89CC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E7AE24F6-004C-4CE3-86D5-58D7F5FE89CC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E7AE24F6-004C-4CE3-86D5-58D7F5FE89CC}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{E7AE24F6-004C-4CE3-86D5-58D7F5FE89CC}.Debug|x64.Build.0 = Debug|Any CPU
+		{E7AE24F6-004C-4CE3-86D5-58D7F5FE89CC}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{E7AE24F6-004C-4CE3-86D5-58D7F5FE89CC}.Debug|x86.Build.0 = Debug|Any CPU
+		{E7AE24F6-004C-4CE3-86D5-58D7F5FE89CC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E7AE24F6-004C-4CE3-86D5-58D7F5FE89CC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E7AE24F6-004C-4CE3-86D5-58D7F5FE89CC}.Release|x64.ActiveCfg = Release|Any CPU
+		{E7AE24F6-004C-4CE3-86D5-58D7F5FE89CC}.Release|x64.Build.0 = Release|Any CPU
+		{E7AE24F6-004C-4CE3-86D5-58D7F5FE89CC}.Release|x86.ActiveCfg = Release|Any CPU
+		{E7AE24F6-004C-4CE3-86D5-58D7F5FE89CC}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -376,6 +404,8 @@ Global
 		{B6D04A38-C980-4CEF-93F7-A8FB80F4EEAB} = {952F4E3A-A0F6-4BBC-9C76-B50EDF490743}
 		{F9B19605-4A41-4ECD-961F-7A7DA5A201D6} = {8F059608-19DA-4E57-8792-C272CA5D3EA6}
 		{040B16E4-BEC3-4429-920D-C0D206B1D6D9} = {F99C02DD-E20B-4564-85FE-C6DB5BCF57A6}
+		{01B12B16-26AB-41EF-9D2F-91894F0CE19A} = {A8A6170C-6952-4985-8317-D765DB05FE47}
+		{E7AE24F6-004C-4CE3-86D5-58D7F5FE89CC} = {A8A6170C-6952-4985-8317-D765DB05FE47}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {8F61C6A5-176D-49B1-99D0-6BDBD8E0168A}
